+++ src/lib.rs
@@ -32,7 +32,7 @@ pub use decimal::{Decimal128, Decimal128BuildError};
 pub use status::{RoundingMode, Status};
 
 #[cfg(feature = "fmt")]
-pub use convert::ParseDecimalError;
+pub use convert::{Engineering, ParseDecimalError};
 
 #[cfg(feature = "binary-float")]
 pub use convert::Decimal128FromFloatError;